Move DownloadPDFButtonWithModal to FormSync folder
