+++ front/app/components/FormBuilder/components/FormBuilderTopBar/messages.ts
@@ -1,10 +1,6 @@
 import { defineMessages } from 'react-intl';
 
 export default defineMessages({
-  downloadPDF: {
-    id: 'app.components.FormBuilder.components.FormBuilderTopBar.downloadPDF',
-    defaultMessage: 'Download as pdf',
-  },
   autosave: {
     id: 'app.components.FormBuilder.components.FormBuilderTopBar.autosave',
     defaultMessage: 'Autosave',
