+++ front/app/components/FormSync/messages.ts
@@ -5,4 +5,8 @@ export default defineMessages({
     id: 'app.components.FormSync.downloadExcelTemplate',
     defaultMessage: 'Download an Excel template',
   },
+  downloadPDF: {
+    id: 'app.components.FormBuilder.components.FormBuilderTopBar.downloadPDF',
+    defaultMessage: 'Download as pdf',
+  },
 });
