+++ front/app/containers/Admin/projects/project/inputImporter/ReviewSection/EmptyState.tsx
@@ -11,7 +11,7 @@ import { useParams } from 'react-router-dom';
 
 import usePhase from 'api/phases/usePhase';
 
-import DownloadPDFButtonWithModal from 'components/FormBuilder/components/FormBuilderTopBar/DownloadPDFButtonWithModal';
+import DownloadPDFButtonWithModal from 'components/FormSync/DownloadPDFButtonWithModal';
 import ExcelDownloadButton from 'components/FormSync/ExcelDownloadButton';
 
 import { FormattedMessage } from 'utils/cl-intl';
